Fraction->double in SwAccessibleMap
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/202402
Tested-by: Jenkins
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/202818
Tested-by: Tomaž Vajngerl <quikee@gmail.com>

+++ sw/source/core/view/pagepreviewlayout.cxx
@@ -1141,7 +1141,7 @@ bool SwPagePreviewLayout::Paint(vcl::RenderContext& rRenderContext, const tools:
         // update at accessibility interface
         mrParentViewShell.Imp()->UpdateAccessiblePreview(
                         maPreviewPages,
-                        aMapMode.GetScaleX(),
+                        double(aMapMode.GetScaleX()),
                         mrLayoutRootFrame.GetPageByPageNum( mnSelectedPageNum ),
                         maWinSize );
     }
